Replicated movement of actors and camera movement
The actor replication now works. The movement is now replicated thanks to an else statement differentitating servers and clients.

The grid size has been increased. A camera and spring arm was added to each player cube and view is swapped to it on movement. This needs polish as the server and client always change dispite not pressing the button themselves.