Movement works and selection works and end game works
Everything so far works!

The movement only goes on highlighted areas (that are set to bMove = true) and highlighting turns areas red up to base_speed distance away.

cubes are destroyed on space bar and once the player number -1 number of cubes are destroyed 'End game" should appear in the logs.