UI interface working plus 1 player start auto
The bluebrint for the UI is up and has the list of actions to be placed. The next step is to list the actions then have an end step that does the actions. As such the begin play randomly selects a starting position, but this only works for one player. Multiplayers have to be selcted individually but it does not crash.