BP pawn
The pawn has been replaced with a blueprint pawn.
This allows for all the bp enabled functions (native and implemented) to work as you are calling the child of pawn. Also the widget can be called from the pawn blueprint instead of level so that way it can be reference directly in the bp pawn and so things like Event Dispatchers can talk to each other.
Currently the move and attack buttons, plus interactions all work without error. Attack is not implemented yet.