Move and attack functions success, maybe
The movement function still jitters but that's because it is defined twice  - on the server and clinet. Without this, however, the clinet will move to the location faster.
Rather than waiting for the movement to finish to "finish it's turn" there's a second flag that all actions turn off.
So far both attack and movement work to the best extent of working.